examples/agent_framework_comparison/requirements.txt
openai==1.52.2
gradio==5.11.0
python-dotenv==1.0.1
arize-phoenix>=4.29.0

langchain==0.2.17
langchain_openai==0.1.25
langgraph==0.2.39
llama_index==0.12.41
llama_index.llms.openai==0.2.15
autogen~=0.2
crewai==0.70.1

openinference-instrumentation==0.1.18
openinference-instrumentation-openai==0.1.14
openinference-instrumentation-langchain==0.1.28
openinference-instrumentation-llama-index==3.0.3
openinference-instrumentation-crewai==0.1.2
openinference-instrumentation-litellm==0.1.4